+++ exporter/otlpexporter/factory.go
@@ -59,7 +59,7 @@ func createTraceExporter(
 	_ context.Context,
 	params component.ExporterCreateParams,
 	cfg configmodels.Exporter,
-) (component.TraceExporter, error) {
+) (component.TracesExporter, error) {
 	oce, err := newExporter(cfg)
 	if err != nil {
 		return nil, err